[PE Helper] DTPE background override support

+++ Helpers/extps1/PE_Helper/PE_Helper.ps1
@@ -606,19 +606,24 @@ function Start-PECustomization
                 icacls "$imagePath\Windows\system32\winpe.jpg" /grant "$(Get-LocalizedUsers -admins $true):(M)" | Out-Host
                 icacls "$imagePath\Windows\system32\winpe.jpg" /grant "$(Get-LocalizedUsers -admins $false):(M)" | Out-Host
                 Write-Host "Changing wallpaper..."
-                switch ($arch)
-                {
-                    x86 {
-                        Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_x86.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
-                    }
-                    amd64 {
-                        Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_amd64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
-                    }
-                    arm64 {
-                        Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_arm64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
-                    }
-                    default {
-                        Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_amd64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                if (Test-Path -Path "$((Get-Location).Path)\backgrounds\wallpaper.jpg" -PathType Leaf) {
+                    # A wallpaper override has been detected. We'll use it instead.
+                    Copy-Item -Path "$((Get-Location).Path)\backgrounds\wallpaper.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                } else {
+                    switch ($arch)
+                    {
+                        x86 {
+                            Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_x86.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                        }
+                        amd64 {
+                            Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_amd64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                        }
+                        arm64 {
+                            Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_arm64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                        }
+                        default {
+                            Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_amd64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                        }
                     }
                 }
                 Write-Host "Wallpaper changed"
